+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 1.22.2 - 2026-02-04
+
+### Changed
+
+- Update project dependencies and configuration
+- New async asset connector base class
+
 ## 1.22.1 - 2026-01-15
 
 ### Fixed